+++ react/components/EXPERIMENTAL_Charts/LineChart/constants.ts
@@ -8,14 +8,11 @@ export const colors = [
   '#0000FF'
 ]
 
-type DefaultLineProps = Pick<LineProps, 'type' | 'strokeWidth' | 'dot'>
 export const defaultProps: DefaultLineProps = {
   type: 'monotone',
   strokeWidth: 3,
   dot: false,
 }
-
-
 
 export const tooltipProps = {
   cursor: false,