+++ react/components/EXPERIMENTAL_Charts/helpers.ts
@@ -1,8 +1,22 @@
+import { merge } from 'lodash'
 import { commonDefaultProps } from './commonProps'
+import { defaultProps } from './LineChart/constants'
 
-const getDefaultProps = (userProps: ChartProps)  => {
-    const configs: ChartProps = {...commonDefaultProps, ...userProps}
-    return { configs }
+const getChartDefaultProps = (userProps: ChartProps)  => {
+  const props = commonDefaultProps
+  userProps && Object.keys(userProps).forEach(key => (
+    props[key] = merge(props[key], userProps[key])
+  ))
+  
+  return { configs: props }
 }
 
-export default getDefaultProps
+const getLineDefaultProps = (userProps: LineProps)  => {
+    const lineConfigs = {...defaultProps, ...userProps}
+    return { lineConfigs }
+}
+
+export {
+  getChartDefaultProps,
+  getLineDefaultProps
+}