+++ react/components/EXPERIMENTAL_Charts/typings.d.ts
@@ -1,9 +1,16 @@
 type BaseGridProps = Pick<CartesianGridProps, 'vertical' | 'horizontal'>
-type BaseAxisProps = Pick<XAxisProps, 'axisLine' | 'tickLine', 'padding'>
+type BaseXAxisProps = Pick<XAxisProps, 'axisLine' | 'tickLine' | 'padding' | 'tick' | 'hide'>
+type BaseYAxisProps = Pick<YAxisProps, 'axisLine' | 'tickLine' | 'padding' | 'tick' | 'hide'>
 type BaseContainerProps = Pick<ResponsiveContainerProps, 'height' | 'width'>
+type DefaultLineProps = Pick<LineProps, 'type' | 'strokeWidth' | 'dot'>
 
 type ChartProps = {
-  axis?: BaseAxisProps,
+  xAxis?: BaseXAxisProps,
+  yAxis?: BaseYAxisProps, 
   container?: BaseContainerProps,
   grid?: BaseGridProps
+}
+
+type LineProps = {
+  type?: LineType
 }